2024.04.22 공통영역, 수정양식 수정

+++ test_jsp/src/main/webapp/WEB-INF/views/board/boardDetailView.jsp
@@ -1,5 +1,7 @@
 <%@ page language="java" contentType="text/html; charset=UTF-8"
     pageEncoding="UTF-8"%>
+<%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core" %>
+
 <!DOCTYPE html>
 <html>
 <head>
@@ -10,22 +12,29 @@
 $(function(){
 	// 삭제
 	$('#deleteBtn').on('click', function(){
-		var bno = '${ board.boardNum}';
-		
-		$.ajax({
-			type: "DELETE",
-			url: "/boards/" + bno,
-			contentType: "application/json; charset=UTF-8",
-			success: function(){
-				alert("삭제완료");
-				location.href = "/board";
-			},
-			error: function(xhr, status, error){
-				alert("code : " + xhr.status + "\n"
-						+ "message : " + xhr.responseText + "\n"
-						+ "error : " + error);
-			}
-		});
+
+		var rlt = confirm('삭제하시겠습니까?');
+
+		if(rlt) {
+			var bno = '${ board.boardNum}';
+
+			$.ajax({
+				type: "DELETE",
+				url: "/boards/" + bno,
+				contentType: "application/json; charset=UTF-8",
+				success: function(){
+					alert("삭제 처리되었습다.");
+					location.href = "/board";
+				},
+				error: function(xhr, status, error){
+					alert("code : " + xhr.status + "\n"
+							+ "message : " + xhr.responseText + "\n"
+							+ "error : " + error);
+				}
+			});
+		} else {
+			return;
+		}
 	});
 
 	// 수정 Form
@@ -37,6 +46,9 @@ $(function(){
 </head>
 <body>
 
+<c:import url="/menubar" />
+<hr>
+
 	<h2>${ board.boardNum } 번 게시글 내용 보기 및 수정 페이지</h2>
 		글제목 : ${ board.boardTitle } <br>
 		작성자 : ${ board.boardWriter } <br>
@@ -44,6 +56,11 @@ $(function(){
 	<br>
 	<button id="modifyBtn">수정하기</button>
 	<button id="deleteBtn">삭제하기</button>
+
+
+
+<hr style="clear:both;">
+<c:import url="/footer" />
 
 </body>
 </html>
